requirements.txt
dash==2.0.0
joblib==1.1.0
pandas==1.3.4
tensorflow-cpu==2.5.0
wordcloud==1.8.1
